requirements.txt
asgiref
blinker
Brotli
certifi
cffi
charset-normalizer
click
colorama
ConfigArgParse
crispy-bootstrap5
cryptography
distlib
Django
django-crispy-forms
django-email
django-environ
django-filter
djangorestframework
djangorestframework_simplejwt
EasyProcess
entrypoint2
filelock
Flask
flask-cors
Flask-Login
gevent
geventhttpclient
greenlet
idna
itsdangerous
Jinja2
locust
Markdown
MarkupSafe
MouseInfo
msgpack
mss
Pillow
platformdirs
psutil
psycopg2
PyAutoGUI
pycparser
PyGetWindow
PyJWT
PyMsgBox
pyperclip
PyRect
pyscreenshot
PyScreeze
pytweening
pytz
pywin32
pyzmq
requests
sqlparse
tzdata
urllib3
virtualenv
Werkzeug
zope.event
zope.interface